7zip is too noisy
As I understand it the developer of 7zip has ignored requests to add a silent switch for some time. This solution is cheesy but will keep things quiet as long as the "Everything is Ok" message doesn't change. If the message changes or 7zip runs into a problem, this should just output the same amount as the existing version, just with slightly deferred output.

Also use `er` instead of `e`, otherwise interrupting a WinRPM install can give some confusing errors regarding 2.71828...